+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Change Log
 ==========
 
+## 1.1.0 SNAPSHOT (2015-07-31)
+
+**Enhancements:**
+
+- \[General\] Increased minimum SDK to 9 (Android 2.3.x Gingerbread)
+- \[GridLayoutManager\] Add support for ```wrap_content```
+- \[AdvancedRecyclerView\] Implement support for OverScrollMode ```ifContentScrolls```
+
+**Bugfixes:**
+
+- \[LinearLayoutManager\] Add item decoration size to size calculation of child view
+
 ## [1.0.2](https://github.com/sandrolutz/AdvancedRecyclerView/tree/1.0.1) (2015-07-27)
 
 **Bugfixes:**